[#7816] Add permission level tests for irepl with groups
Tests for irepl with various permission levels is already captured
in test_irepl.test_all_permission_levels__issue_7444_7465. However,
this test suite is only capturing permission levels for users. The
tests do not capture the situation in which a user has no permissions
on a data object but is a member of a group which has permissions on
the data object.

This change modifies the existing tests such that the existing tests
are preserved but also allow for modifying the permissions for groups
as well.