[#7816] Check group permissions for replication
In addition to checking whether a user has permission to replicate
a data object, all groups of which the user is a member should be
checked as well. If either the user or any of the user's groups have
sufficient permission to replicate the data object, then the replication
should be allowed.