[#7725] test_irodsctl: Adjust assertions for delay server leader
If the delay server leader is the local server running the test, there
should be at least 3 processes reported by the library function called
capture_process_tree. If the delay server leader is NOT the local server
running the test, there should be at least 2 processes reported. The
likelihood of there being more than 2 processes in the case where the
local server is not the delay server leader is not very high which can
lead to occasional failures unrelated to the issue being tested.

The assertions for the test called
test_agents_kept_alive_by_connected_clients_are_cleaned_up_after_failed_graceful_shutdown__issue_7619
have been adjusted to account for these cases, and extensive explanatory
comments have been added to the code to explain why the test changes
based on this condition.