[irods/irods#7760] Remove rulemsiServerMonPerf.r
This rule file is no longer needed by the test generator because a
test specifically for this microservice has replaced it. The specific
test has been introduced because extra setup is required to run tests
for this microservice.

This is the name of the test which replaces the one which would be
generated by the file which is being removed in this commit:
test_all_rules.Test_AllRules.test_msiServerMonPerf_default_3736