[#5593] Fix test_iget/irepl_with_purgec assumptions
The assertions for some implementations of test_iget_with_purgec and
test_irepl_with_purgec are assuming that the first cache replica is
always selected for read and that therefore the first cache replica
will always be purged. This will not necessarily be the case. The
test has been updated such that the assertions detect which cache
replica was purged and assert that the other cache replica was not
purged.