[#7859] Convert null pointer to empty string for checksum output parameter of msiDataObjChksum.
Before this commit, the following conditions could lead to a segfault.

    - The python rule engine plugin is enabled
    - A rule written in python invokes msiDataObjChksum
    - Checksum verification is performed and no issues are found

The segfault is the result of the microservice setting the checksum
output parameter to a null pointer (because it did not find any issues)
which eventually leads to iRODS attempting to construct a std::string
with the null pointer.

The construction of the std::string occurs in convertFromMsParam.
Passing a null pointer to the std::string constructor is undefined
behavior.