[#7918] replica_truncate: Add unit test output string checking
This change adds assertions about the contents of the output string
that is returned from the replica_truncate API. Depending on the error,
the message and other contents of the JSON structure could be in a
variety of states, so we want to make sure that these hold the expected
values in all cases. In the cases where errors do not occur, we want to
ensure that the correct replica number and resource hierarchy are returned
so that the caller knows which replica was truncated.

This change also adds a new test for when a logical path including a zone
name which does not exist is targeted for truncate.