[#8036] Communicate server status to service manager
In addition to the service manager communication functionality, this commit
introduces `distro_distill.py`, a slightly updated version of the
`distro_info.py` module currently in the externals repo. The goal is to
eventually make a standalone package for `distro_distill.py`. The buildsystem
now uses this instead of the `distro` Python module or `lsb_release` to
populate the `IRODS_LINUX_DISTRIBUTION_*` variables. Downstream projects may
need to be updated accordingly. As this results in potential changes to cache
variables, a clean build may be desirable.

This commit also introduces the header `irods/chrono.hpp` in `lib/core` which
contains a couple of funcitons related to timestamp generation. The code for
getting the monotonic timestamp for service manager communication has enough
in common with the timestamp string generation code used by the logger that it
made sense to consolidate them both into the same header.

Also:
  - [#8123] Package systemd service unit template
  - [#8125] Add `procps` to recommends for server deb package