[#7722] Validate and reject invalid zone names.
Before this commit, it was possible for an administrator to change the
zone name of an existing zone to an invalid name. The reason for this is
because the rsGeneralAdmin API did not validate the new zone name.

This commit makes it so that the server validates the new zone name
before updating the catalog. If the new zone name does not satisfy the
regex, `^[A-Za-z0-9_\.]+$`, it is rejected.

This commit also makes the server use the same validation logic for
additions of new remote zones. This makes it easier for administrators
to spot zone name validation errors in the log file.