[#2978] GeneralAdmin: Deprecate "modify" of "user" of type "rodsgroup"
This commit deprecates using the "modify" operation on a "user"
when the user is of the "rodsgroup" type. The deprecation comes
in the form of a warning message from the API put on the rError
stack. This will be replaced with an error in the next major release.

Updates all of the tests in test_iadmin.test_moduser_group such that
the deprecation warning message is expected on valid group modifications
using moduser.