[#4984] Fix unencrypted parallel transfer with server encryption enabled
This change fixes an agent crash caused by a SIGABRT due to
uncaught exceptions when a client using unencrypted communications
with one server attempts to put or get data via legacy parallel
transfer to another server which is using encryption in its
server-to-server communications. The change has two pieces:

1. When the parallel transfer portal is being set up in rsDataGet
   and rsDataPut, the "negotiation results" of the client-server
   connection are captured in the conditional input of the portalOpr
   in order to later determine whether encryption ought to be used
   in communications via the direct legacy parallel transfer portal.

2. The existing funciton used in the data-moving threads for legacy
   parallel transfer has been wrapped such that any exceptions it
   may throw will be caught in order to prevent the agent from
   aborting.

In this way, even if the determination of whether to use encryption
in portal communications is incorrect, the exceptions being thrown
as a result will at least be caught and handled.