[#3581] Avoid SIGABRT by setting pointers to null after deallocation.
Before this commit, irule could cause the agent servicing the request to
trigger a SIGABRT if invoked under the correct conditions. This happened
because the iRODS Rule Language would deallocate memory, but not set the
pointers to null afterwards. While the client would receive the correct
error code, the agent would crash during tear down because it would
attempt to deallocate memory that was not allocated. This occurred
during the call to clearMsParam().