[#7412,#7491] Remove test_exception_in_delay_server
This test is meant to demonstrate that the delay server
does not terminate when an exception is thrown and that
delay rules continue to execute. The delay server's
resilience in the face of exceptions as well as machinery
around ensuring the delay server is running (via delay
server migration) is much better now.

Additionally, this test is causing problems entirely
unrelated to the behavior it seeks to demonstrate, so we
have decided to remove it.