[#6766] ModAVUMetadata: Remove rmw
Wildcard support for removing metadata AVUs remains in
the database plugin due to its usage in some internal
implementations, such as the "set" operation of
ModAVUMetadata. The breaking change here is the removal
of the "rmw" operation from the API. The internal
implementation of wildcard support can be removed at any
time as it is no longer directly accessible to external
interfaces.