[#8265] Restore support for KeyValPair in Python REP
It is common in iRODS for the Python REP to allocate KeyValPairs using
MSIs provided by the REPF. However, iRODS 4.3.4 accidentally broke
support of that in an attempt to address a memory leak.

This commit restores support for KeyValPairs by reverting the fix for
the memory leak. This isn't ideal, but breaking support of KeyValPairs
in the Python REP is much worse.