[8265] Restore support for KeyValPair in Python REP.
In iRODS 4.3.3, it was common to pass KeyValPair's between the iRODS
Rule Language and Python REP. However, iRODS 4.3.4 accidentally broke
support of that in an attempt to address a memory leak.

This commit restores support of KeyValPair's by reverting the fix for
the memory leak. This isn't ideal, but breaking support of KeyValPair's
in the Python REP is much worse.