[ 7483] Use OpenSSL 3.0 in sslSockComm and SSL plugin
This commit also removes the "fallback" DH parameters historically
used in the absence of a DH parameters file. This behavior causes
many warning messages to appear in the log despite client-server
communications working. This indicates that the DH parameters file
is the preferred usage, over the pre-generated DH parameters
hard-coded in this open source project.