[#8261] GenQuery2: Align handling of DISTINCT keyword with SQL standard.
This commits makes it so that the handling of the SQL DISTINCT keyword
mirrors that of the SQL standard. The decision for when to include the
DISTINCT keyword is left to the caller. As a result, the NO DISTINCT
syntax has been removed.

This is primarily about making the SQL generated by the parser more
deterministic.