[8261] GenQuery2: Allow SQL DISTINCT in functions.
This commit makes it so that users can specify the DISTINCT keyword in
function calls. This is primarily about fixing support of count-based
queries. The following syntax is now valid:

  "select count(distinct C1, C2, ...CN)"

While the parser will accept the DISTINCT keyword for all function
calls, the underlying RDMS is not guaranteed to support the generated
SQL.