[#8016] Enable auth plugins to accept passwords in requests
This commit allows the AUTH_PASSWORD_KEY to be used to provide a
password directly to the authentication plugins in order to bypass
prompts and input from stdin as well as reading the password from
an existing irodsA file. The priority order of determining the
password to use for authentication has been changed. If any of the
following conditions is not met, the plugin will try the next one
until it reaches the end:

1. If the anonymous user is authenticating, no password is used.
2. If the client has set the "force_password_prompt" option, prompt
the user for a password from stdin. This is how iinit always gets a
password prompt to appear.
3. If the AUTH_PASSWORD_KEY is present, use the password provided therein.
4. If a password can be read from an existing irodsA file, use that.
5. Finally, if none of the above are true, prompt the user for a password
from stdin because no passwords are available any other way.

Cherry-picked from: 62bd385c741541df32f836e952f04748e6db0181