[#8257] Move writing irodsA file to native auth plugin
This commit deprecates clientLoginTTL. Its functionality has been moved
into the native authentication plugin as described below.

This commit adds two new operations to the native authentication plugin:
record_authentication_file and complete_authentication.

record_authentication_file is an extra step added to the plugin to write
the irodsA file after successfully authenticating the user. If valid TTL
information was provided by the caller, the functionality of clientLoginTTL
is invoked from the plugin. A limited password is generated for the user
in the server,

complete_authentication is mainly a formality for clarity in the flow of
plugin operations. It sets the loggedIn flag and marks the authentication
flow as complete.

This commit adds a new parameter recognized by the native authentication
plugin called "record_auth_file". This allows the caller to control
whether the record_authentication_file step is performed. It defaults to
false if not provided. If this option is not set or is set to false, the
irodsA file is not written by the plugin (even if a TTL parameter was
provided!).