[#7948] Fix iinit test expecting user password prompt
iinit itself only prompts for the user's password for native auth
scheme now. The other schemes' plugins are responsible for getting
the user's password, so if something causes the connection with the
server to fail or load the plugin, a password prompt will not be
shown. This commit adjusts expectations for a test with this
situation.