[#8360] pam_password: Return error if TLS not enabled
The client-side plugin will return an error if TLS is not enabled
because if it reaches the server, the sensitive information will
have already been sent over an insecure channel. However, the server
also returns an error just in case a different client tries to get
around the TLS requirement.