[#8360] Remove test_authentication_PAM_without_negotiation
This test is no longer necessary because negotiation is always
enabled now and TLS is required for pam_password. There is a new
test for no-TLS pam_password authentication called
test_user_cannot_authenticate_with_CS_NEG_REFUSE_even_if_tls_is_configured__issue_8360
which acts as a replacement for this test.

Also, replaced a condition to fail a test in favor of raising an
error when a precondition is not met.