[#8147] Make agent factory notify main server process of readiness
This commit makes it so that the agent factory notifies the main server
process when it has completed initialization. Initialization of the
agent factory is considered complete after it has opened its listening
socket and is entering its main processing loop.

This behavior is achieved through the use of the SIGUSR2 signal.