[#8373] DataObjGet: Enforce max size for single buffer
The maximum size for a single buffer get is 2047 MB because 2048 MB
causes an overflow in 32-bit signed integers. This change ensures that
the max single buffer size is in the range [1, 2047] to prevent issues
around allocating the buffer.

This commit also ensures that the buffer allocated for single buffer
gets is properly cleared.