[#7229] Run plugin start operations before REPF interactions
This commit addresses an order-of-operations issue in which the REPF was
invoking the `rule_exists` plugin operation before the `start` plugin
operation.

This commit makes it so that agent initialization runs the `start`
plugin operation before any PEPs are triggered. This guarantees that
rule engine plugins are in a usable state before the REPF attempts to
interact with them.

Resolves the following additional issues:
- https://github.com/irods/irods_rule_engine_plugin_python/issues/241
- https://github.com/irods/irods_rule_engine_plugin_python/issues/242
- https://github.com/irods/irods_rule_engine_plugin_python/issues/243
- https://github.com/irods/irods_rule_engine_plugin_python/issues/244