[#8529] Return SYS_EXCEED_CONNECT_CNT when server redirects to self due to misconfiguration
This commit restores the server's ability to detect infinite loops in
server redirection.

This situation occurs when the connected iRODS server has multiple
names, but only recognizes a subset of them.

This creates a situation in which the iRODS server attempts to redirect
the request to the "correct" computer, but the "correct" computer turns
out to be itself. This results in the endless creation of iRODS agents.