[#8586][irods/irods_resource_plugin_s3#2146] Calling fileNotify on close
This is to fix S3 issue 2146. The issue occurs on large file transfers
when the resource is in attached and decoupled modes and a redirect is
required.  The following is the rough description assuming the client
is connected to server A and the resource is attached to server B.

 1. s3_resolve_resc_hier_operation is run on server A.
 2. The db_reg_data_obj_op operation is called on the catalog provider
    to create the data object entry. After this there is a record in
    the database for the replica.
 3. The request is redirected to server B.
 4. On server B, in s3_file_create_operation the physical path is paved
    over.  This allows the plugin to write to the correct S3 key.
 5. If the file is a small file, s3_notify_operation is called on server
    A. The plugin uses this to pave over the physical path within server
    A's agent memory.
 6. If the file is a large file, the notification operation is not
    called on server A. The plugin has no opportunity to pave over the
    physical path within server A's agent memory.
 7. The file is written to S3 (from server B).

There is no opportunity to update the physical path on server A and
therefore the database never gets updated.

This fix is to call fileNotify on the close. This gives the S3 resource
the opportunity to overwrite the physical path in memory which gets
represented back in the database.

Notes:
1. The object ID (reversed) is used as the path in decoupled mode. This
   is done as it provides a deterministic unique path.
2. I considered a hash of the path in iRODS but this is not guaranteed
   to be unique as a file can be put and moved and a new file put to
   this path.
3. A path plus date is not deterministic as the date could change
   between server A and server B paving over the path, and is also not
    guaranteed to be unique if a move happens in the same day.
4. The problem with the object ID is that it doesn't get created until
   step 2, and for large file uploads step 1 is the only opportunity to
   pave over the path on server A.

This solution seems to check all of the boxes with a simple extra call
to fileNotify. This is a NOOP for all other resource types.