[#8384] Add additional checksum return checks to rsync
This will check for CHECK_VERIFICATION_RESULTS from rsDataObjChksum and then
parse the resultant JSON array. Originally, CHECK_VERIFICATION_RESULTS was
erroneously treated as an error and irsync would terminate abnormally because
of it. Now, if any error_code is nonzero, it will still fail and return while
logging a message (i.e. all 0 error codes indicates everything is fine). For
the case listed in the ticket, the resultant JSON has a 0 error code: it's
informing the caller that the replica is skipped due to it being stale.
Therefore, it will pass and allow the copy to progress.