[#8554] Define/call RESOURCE_OP_READ_CHECKSUM_FROM_STORAGE_DEVICE.
The inputs (other than plugin_context) to this operation are:

- _checksum_scheme (INPUT) - The desired checksum/hash scheme.
- _returned_checksum (OUTPUT) - The returned checksum if it is
  available.

When a checksum is required on an object, first this operation is
called on the storage resource to determine if the desired checksum
is directly available. If it is, the storage device returns the checksum
and the full file read / checksum calculation is short circuited.

If the desired checksum is not avaiable or this operation is not defined for
the storage plugin, the full file read for checksum calculation is
triggered.