[8758] Refactor quota query so quotas are properly processed
Previously, the quota system did not function properly due to several
misconceptions. The queries for fetching quota rows returned them
in most-severe order, but the linked list created from the rows
reversed it, so quota code that only read the head of the list only
processed the least-severe quota. As a result, after rectifying the
queries, the two-pass quota check in rsGetRescQuota is no longer
needed as a single query properly handles both total and resource
quotas.