[#8756][#8773] Skip CRC64/NVME tests while in topology.
These tests change the default_hash_scheme to "crc64nvme" in
server_config.json on the server that is running the test. In some
configurations the actual resource written to is not the same as
the server where the test is running. This means that the hash
scheme won't match what the test is expecting.

Since we have no way to update the server configuration on all servers,
these tests will now be skipped when running in topology.